Use i64 number (#112)

+++ vm/src/value.rs
@@ -31,7 +31,7 @@ impl Value {
         if self.is_cons() {
             TypedValue::Cons(Cons::from_raw(((self.0 as i64) >> 1) as u64))
         } else {
-            TypedValue::Number(Number::new(self.0 >> 1))
+            TypedValue::Number(Number::new((self.0 as i64) >> 1))
         }
     }
